+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 
 ### Features
 
-- Expose Android screenshot masking options (`screenshot.maskAllText`, `screenshot.maskAllImages`) for error screenshots ([#6007](https://github.com/getsentry/sentry-react-native/pull/6007))
+- Expose screenshot masking options (`screenshot.maskAllText`, `screenshot.maskAllImages`) for error screenshots ([#6007](https://github.com/getsentry/sentry-react-native/pull/6007))
 
 ## 8.8.0